Added forgotten file to previous commit

+++ bookifySystem/src/main/java/com/bookify/search/SearchPreviewDTO.java
@@ -2,4 +2,4 @@ package com.bookify.search;
 
 //TODO: add thumbnail guid, cost per day, room type
 public record SearchPreviewDTO(int roomID, String name, float rating, int reviewCount,
-                               int bedCount, int price, String roomType, String thumbnail) {}
+                               int bedCount, float price, String roomType, String thumbnail) {}
